Tests/LibWeb/Text/expected/svg-relayout-relative-inset-stable.txt

before: 30,20
after first relayout: 30,20
after second relayout: 30,20
